think: Organize workspace internals
Keep workspace types, state compatibility, and provider entrypoints together. Remove the unused generated adapter import and document the deprecated state compatibility bridge.
